Add iOS community triplets and non-macOS (iOS) toolchain support
The toolchain is automatically picked up when CMAKE_SYSTEM_NAME
equals iOS. The toolchain selects the architectures for which to build
packages, as well as the processor type (because certain packages like
libpng depend on the processor type value).

Automatic detection of triplets for iOS / tvOS / watchOS in the
vcpkg.cmake toolchain when building your own application.
It uses the values of CMAKE_SYSTEM_NAME and CMAKE_OSX_ARCHITECTURES
to detect the triplet.

Added workaround so find_package finds vcpkg installed packages
when targeting iOS.
This is done by saving / restoring the value of CMAKE_FIND_ROOT_PATH
while also adding the vcpkg package root in the find_package override
macro.
The workaround can be removed once vcpkg upgrades to CMake 3.15.0
or higher where the issue is fixed.

Fixes: #6003